separated page and render pipe code
